Revert "Shutdown plugins on window close"
This reverts commit f8a6d31af565bf0a316de7c8cd0ddf5dee9205f2.

This never worked because it ran stopPlugin on the AWT thread
which calls invokeAndWait, which throws when not on the awt
thread. Also stopPlugin only attempts to stop plugins that are
marked as disabled, but it is only called for plugins which are
enabled, so nothing actually happens. This also creates a deadlock
if you try to close the client while starting it, because
startPlugin is locked on one thread, but stop plugin is trying
to take the lock on the edt.