raids plugin: fix matching rotation whitelist
Spaces were not being removed from the config values causing the
whitelist contains check to fail.

Additionally, change the format of the raids whitelist to be newline
separated CSV which is not interpreted by GSON as valid JSON, fixing the
config service from serializing the whitelist rotation which is adding
erroneous quotes around the room names.

This also refactors the rotation whitelist display to be a simple yes/no
instead of this unused match count logic.

Co-authored-by: Alexsuperfly <alexcsumner@gmail.com>