swingutil: Set button tooltip text in addModalTooltip
Prior to this commit, addModalTooltip only added a button listener to
update the tooltip when its state changed. This, however, did not set
the initial tooltip and leaves modal tooltip buttons in the client
lacking a tooltip until they are toggled. This commit sets the initial
tooltip state using the same toggle check as is used in the listener.