Notifier: Reuse Clip instances
This was leaking instances of Clip, which on linux/alsa+alsa-pulseaudio
keeps a connection to pulseaudio open forever, for each notification
which will eventually lock/crash the pulse daemon.

If we just made it close the clip, it would become difficult to change
the volume because the volume interface would go away as soon as the
clip has stopped playing, so instead we keep the clip around after it
has been loaded if the mtime of the file hasn't changed.