tagtabs: split tag search from bank search
This fixes a lot of edge cases which required us to re-search for things
after input was re-used when a tag tab was open, and removes the
requirement of having the bank search text be in sync with the
currently selected tab.

This also no longer parses the tag tab out of the title during times
when the input is being re-used (like Withdraw-x) which was broken
earlier from the bank plugin changing the title to include the value of
searches.

A new tag tab title is being set now on all tag tab views so the old tag
tab tab title event can be removed.