screenshot: Fix BA high gamble toggle
Because both level up dialogs and BA high gambles use the same widget
interface, commit ec75d58ee968fde5ecfb40a0a35df93407aed5cd caused BA
high gambles and some level up dialogs to trigger screenshots if either
config was enabled in the config, as that would set a flag to capture a
screenshot on the next game tick based on the widget text. This commit
adds config checks to the DIALOG_SPRITE_TEXT branch to ensure the
appropriate config flag must be set for a screenshot to be captured.