gpu: use noperspective for hsl interpolation
Prior to 9949629b9367b25211715038b8fe0c38d8753f41 the interpolation of fragment
shader inputs appears to not have been accounting for perspective. This broke
texturing due to the texture ids being interpolated for perspective which was
fixed in 6d28da9c0a8cbdc13647d23c020bd7d55a715a43. The hsl values similarly
require not accounting for perspective, so the noperspective interpolation
qualifier has been added. It is unclear to me why the geometry shader removal
would affect the interpolation like this since it was emitting vertices with
the correct z values.