client: rename jogl/gluegen osx natives artifacts
gradle does not look at classifiers when doing dependency conflict
resolution, causing it to not correctly handle having multiple
dependencies of differing versions and classifiers. It picks the highest
version of any of the dependencies with the same groupId and artifactId
and uses that for every other dependency regardless of classifier.

This renames the osx natives artifacts to have a unique artifact name to
workaround this issue