Fix race with client startup and low detail/chat icon loading
Since e9913d7d7d052707927434ad8f86c3debcc7cf66 moved client startup to
be much earlier, the plugins can now start after the client is already
at the login screen. This causes the low detail plugin and chat icon
manager to not initialize correctly as they are waiting on the event
for the login screen.