clientui: setResizable after setVisible
Prevents an issue with clients being resizable with toolbar-less means in some environments (e.g. double clicking title bar on macos).

Honestly I'm not entirely sure why this is ignored if the window isn't visible, rather than being propagated to the window manager when the window is shown, but I guess Swing isn't that smart (like always).

I originally found the information on [this SO answer](https://stackoverflow.com/a/14884056) and it seemed to resolve the issue.

Closes #14235