gpu: use floats for projection on cpu
The use of vanilla's projection for face sorting and back culling mixed
with the plugin's projection matrix causes subtle visual discrepancies,
most commonly noticed on the GE floor and also on player faces.

This moves the face sorting logic to the plugin and uses floats for
projecting the verticies for face sorting, removing these discrepancies.