keybind: relax modifier matching when checking key released
The existing code checks only the keycode for KEY_RELEASED to allow
releaseing keys in any order (such as key releaseing ctrl and then a
when triggering a ctrl+a keybind).

This extends the checks for modifier-only keybinds to allow other
modifiers to be present at the time of the key release. Check all of the
required modifiers were pressed, and the key being released is one of
them.