gpu: fix amd miscompile of frag.glsl
amd drivers 22.7.1 and later miscompile a switch with two leading cases when one falls through.

The switch was originally put in place due to intel drivers breaking with a chain of ifs, which was fixed in 8ca07b65b950a95859e1be55180cf643be79c655.

Hopefully this works better since there is an else {} case and also the color blind code is not behind a condition anymore.