screenshot: fix scaling on macos
Mac OS automatically scales the fbo with gpu on, which is why the gpu
plugin has the macos checks to disable its scaling code. Due to this,
the glReadPixels() call returns an unscaled image.

Revert to the previous behavior on macos of not scaling either the
non-gpu screenshot or the final image with include client frame on.

Another option is to blit and upscale the image when reading it off of
the GPU. This does work, but is a bit more complex:

	int scaledFbo = GL43C.glGenFramebuffers();
	GL43C.glBindFramebuffer(GL43C.GL_FRAMEBUFFER, scaledFbo);
	int scaledRbo = GL43C.glGenRenderbuffers();
	GL43C.glBindRenderbuffer(GL43C.GL_RENDERBUFFER, scaledRbo);
	GL43C.glRenderbufferStorage(GL43C.GL_RENDERBUFFER, GL43C.GL_RGBA, scaledWidth, scaledHeight);
	GL43C.glFramebufferRenderbuffer(GL43C.GL_FRAMEBUFFER, GL43C.GL_COLOR_ATTACHMENT0, GL43C.GL_RENDERBUFFER, scaledRbo);

	int status = GL43C.glCheckFramebufferStatus(GL43C.GL_FRAMEBUFFER);
	if (status != GL43C.GL_FRAMEBUFFER_COMPLETE)
	{
		throw new RuntimeException("fbo incomplete");
	}

	GL43C.glBindFramebuffer(GL43C.GL_READ_FRAMEBUFFER, awtContext.getFramebuffer(false));
	GL43C.glBindFramebuffer(GL43C.GL_DRAW_FRAMEBUFFER, scaledFbo);
	GL43C.glBlitFramebuffer(0, 0, width, height, 0, 0, scaledWidth, scaledHeight,
		GL43C.GL_COLOR_BUFFER_BIT, GL43C.GL_NEAREST);