config: Add default profile linking to config profiles
This commit adds the ability to mark config profiles as being the
default for a set of rsprofiles, meaning that profiles will change
automatically on rsprofile change. This will likely resolve most cases
for players wanting to use the `--profile` launcher argument, and make
it easier to use game mode-specific, or account-specific settings, as it
can be made automatic.