gpu: remove sin/cos table from UBO
This was added in 30513fe55c7bd7dcc6f2e6112b8f9e1f936b9966 due to
inconsistencies in the computed sin/cos values causing graphical issues
on some cards. However the sin/cos table is no longer used for the
projection, and instead only for computing face distance for face
sorting, so I think it is safe to remove.