screenshot: image capture: support insets on client frame
The insets include the native client border, which results in an 8px
alpha border on the image. Remove the insets from the screenshot.

With j2d scaling the image size is an extra pixel wide and the client
position within the frame is an extra pixel down (with gpu). I think
this is due to the rounding of getScaledValue(), which was originally
copied from jogl.