bank tags: use only dynamic components for tab layer
Static components are unsafe due to it causing component ids to the new
static components to be stored and referenced by the client later, like
in the menu, which does not do bounds checking.

This fixes a crash where you right click open the menu on a tag tab,
close the bank, then click the menu.