gpu: fix priority face average calculation
While the client defaults the averages to 0, it also offsets the
distances by the model radius, meaning 0 is always the closest distance.

This is a regression from 25621e9e7b43b25357f4c6273454d031ad4f0b35 which
removed the radius offset from the distances.

Initialize averages to -6000 because the model max diameter is 6000 and
so it should always be less than the computed distance.