timers: Remove sire stun timer on walking sire spawn
When killing either of the southern respiratory systems during the first
phase of the Abyssal sire fight last, sire can change to his walking
form while being far enough from the player that the NpcChanged event
would not be fired. To handle this case, we can also clear the sire stun
timer when that form spawns (when the player walks north, or when sire
walks south).